Splitting the matplotlib files into multiple
 - The goal is to make the API much more consistent by splitting it into different files.
 - The next goal would be to create a class MATPLOTLIBGRAPHIC and to generate the graphic from this class instead of having multiple different functions with redundant code.
