+++ verticapy/plotting/_matplotlib/__init__.py
@@ -14,4 +14,12 @@ OR CONDITIONS OF ANY KIND, either express or implied.
 See the  License for the specific  language governing
 permissions and limitations under the License.
 """
-from verticapy.plotting._matplotlib.plot import *
+from verticapy.plotting._matplotlib.animated import animated_bar, animated_bubble_plot, animated_ts_plot
+from verticapy.plotting._matplotlib.bar import bar, bar2D, hist, hist2D, multiple_hist
+from verticapy.plotting._matplotlib.core import compute_plot_variables, gen_cmap, gen_colors
+from verticapy.plotting._matplotlib.boxplot import boxplot, boxplot2D
+from verticapy.plotting._matplotlib.heatmap import cmatrix, contour_plot, hexbin, pivot_table
+from verticapy.plotting._matplotlib.pie import nested_pie, pie
+from verticapy.plotting._matplotlib.scatter import bubble, outliers_contour_plot, scatter_matrix, scatter
+from verticapy.plotting._matplotlib.spider import spider
+from verticapy.plotting._matplotlib.timeseries import acf_plot, multi_ts_plot, range_curve, ts_plot
